Port the campaign battle routes and battle_stats lineup view
Port campaign.py's battle routes into Campaign.kt (decode/seed/checks, drops
SplitMix64, settle_win + grant/achievement/pinned-exp/flush, plan_battle /
plan_auto / plan_reentry / plan_star_box, helper participant + lineup, first
kill) and the two battle_stats additions (battle_actors, lineup_view_payload).

battle_seed returns the unsigned u64 (BigInteger) the reference stores in the
transaction detail; the route feeds its bit pattern to the drops PRNG as a
Long.

+++ server-core/src/main/kotlin/io/github/okexodus/openknights/server/session/Session.kt
@@ -1549,11 +1549,11 @@ class Session(val service: Service, val kind: String, private val gamePort: Int)
                         "campaign_battle" -> {
                             val helper = Campaign.helperParticipant(request, owned, worldCtx, social, now, inputs)
                             helperRole = helper?.participantId
-                            p = Campaign.planBattle(request, owned, cur, inputs, document, now, world, helper, helperState(helper), seed)
+                            p = Campaign.planBattle(request, owned, cur, inputs, document, now, world, helper, helperState(helper), seed.toLong())
                             if (regen != null) p.packets = listOf(regen) + p.packets
                             if (!Py.truthy(p["commit"])) throw Campaign.Lost(p)
                         }
-                        "campaign_auto" -> p = Campaign.planAuto(request, owned, cur, inputs, document, now, Campaign.SplitMix64(seed xor 0xA070L))
+                        "campaign_auto" -> p = Campaign.planAuto(request, owned, cur, inputs, document, now, Campaign.SplitMix64(seed.toLong() xor 0xA070L))
                         "campaign_reentry" -> p = Campaign.planReentry(request, owned, inputs, document, now, servedTime(cur, now))
                         else -> p = Campaign.planStarBox(request, owned, inputs, document)
                     }
